+++ local_examples/05_graphics_sampling.R
@@ -1,3 +1,5 @@
+# required libraries
+library(AirSensor)
 
 pat <- example_pat